Add Router logic Clean up ~~Summary~~ Home Page
