+++ src/sections/Shard/index.css
@@ -1,4 +1,3 @@
-#SummaryPage,
 #ShardLandEndCountdown {
   display: flex;
   flex-flow: column nowrap;
